refactor(swarm_team): move fallback agent & template to agent planner
